The assembler is dumb about complicated immediates.
For a statement like this:
   $r3 <- 1+4
the assembler would parse it up to the '+' sign,
determine it being a binary operator, parsing the
4 and realizing that it's two immediate arguments,
which the instruction '+' doesn't support.

So, instead of trying to fix the assembler (for now),
we simply put  parens around the expression, which
forces AS to parse the line correctly.