Trampoline code was busted. Now it's not.
The problem here was that the dynamic code referenced the static
location for the data fields. The updates
happened to the dynamic copy and thus
didn't reflect in the actual execution.

The current implementation instead
uses immediate loads and the immediates
are patched up in the dynamic copy.

Theoretically $pc-based loads could also be used,
which would be similar to the original intent, but
1. Current assembler doesn't like it (which is a bug,
   but not worth fixing because of (2).
2. The new ISA, where $pc is not part of the general register bank
   won't support $pc-based indirect loads anyway.