Fixed struct output ptr interference with thunks.
I'm not that happy with the solution because it
requires yet another call-clobbered register ($r10),
but it at least stabilizes the registers used for true
arguments.

Without this, $r4 might be the structure value address
and the first argument can be either $r4 or $r5, depending
on the return value.

I fear that this would wrack havoc with vaargs not to mention
all the other cases, where it would be hard to figure out
which register holds what upon  a function call.

For what it's worth ARM(32) uses this invisible first
argument scheme (AARCH64 went away from it), so
there might be some examples of what needs to happen
there.